more debugging

+++ aspire/preprocessor/downsample.py
@@ -26,7 +26,7 @@ def downsample(stack, n, mask=None, stack_in_fourier=False):
     size_out = np.square(n)
     mask = 1 if mask is None else mask
     num_images = stack.shape[0]
-    output = np.zeros((num_images, n, n), dtype='float32')
+    output = np.zeros((num_images, n, n), dtype='float64')
     images_batches = np.array_split(np.arange(num_images), 500)
     for batch in images_batches:
         if batch.size:
